hubsign/requirements.txt
# Django
Django>=4.2,<5.0
django-environ>=0.11.2

# Django REST Framework
djangorestframework>=3.14.0

# Security
django-cors-headers>=4.3.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
django-debug-toolbar>=4.2.0
